Fix dmenu desktop patch
The path for the .desktop files was hardcoded, and didn't work on NixOS.
I changed it so that it now reads the XDG_DATA_DIRS env var.

Also the .desktop files wouldn't get executed properly by xdg-open for
some reason, so I had to extract the exec command from the files and run
it manually against the selected desktop entry's path.